+++ src/app/feat-recruit/interactions/buttons/send_notify_to_host.ts
@@ -18,8 +18,8 @@ import { searchAPIMemberById } from '../../../common/manager/member_manager';
 import { searchMessageById } from '../../../common/manager/message_manager';
 import { exists, sleep } from '../../../common/others';
 import { sendErrorLogs } from '../../../logs/error/send_error_logs';
-import { joinRequestConfirmButtons } from '../../buttons/create_join_request_buttons';
 import { messageLinkButtons } from '../../create_recruit/common/create_recruit_buttons';
+import { joinRequestConfirmButtons } from '../../join_request/create_join_request_buttons';
 
 const logger = log4js_obj.getLogger('recruitButton');
